Restructure components lib
