+++ src/components/forms/SolutionForm/Title.js
@@ -2,10 +2,10 @@
 
 import { PureComponent } from 'react'
 
-import elem from '../../utils/elem'
+import elem from '../../../utils/elem'
 
-import theme, { breakpoints } from '../../styles/theme'
-import typo from '../../styles/typo'
+import theme, { breakpoints } from '../../../styles/theme'
+import typo from '../../../styles/typo'
 
 const cmz = require('cmz')
