+++ src/components/ui/ApplicantBadge.md
