+++ src/components/ui/CollapsibleSection.md
