+++ src/components/ui/RoadmapHero.md
@@ -13,6 +13,6 @@ RoadmapHero example with image url in props:
 <RoadmapHero
   heading="Next Steps"
   content="Thanks for completing the challenge and telling us a bit about yourself. Next, you will soon receive an e-mail from an X-Team Ambassador who will help us determine if you are a good fit for X-Team."
-  imgUrl={require('../assets/next-steps.png')}
+  imgUrl={require('../../assets/next-steps.png')}
 />
 ```
