+++ src/components/ui/RoadmapTimelineElement.js
@@ -1,7 +1,8 @@
 // @flow
 
 import React, { PureComponent } from 'react'
-import theme from '../styles/theme'
+
+import theme from '../../styles/theme'
 
 const cmz = require('cmz')
