+++ src/index.js
@@ -1,25 +1,25 @@
-import ApplicantBadge from './components/ApplicantBadge'
-import ApplicantListFilter from './components/ApplicantListFilter'
-import Button from './components/Button'
-import CollapsibleSection from './components/CollapsibleSection'
-import ColorPalette from './components/ColorPalette'
-import ErrorBox from './components/ErrorBox'
-import HeaderBar from './components/HeaderBar'
-import HorizontalRuler from './components/HorizontalRuler'
-import Milestones from './components/Milestones'
-import Roadmap from './components/Roadmap'
-import RoadmapHero from './components/RoadmapHero'
-import RoadmapLevel from './components/RoadmapLevel'
-import RoadmapTimelineElement from './components/RoadmapTimelineElement'
-import SvgIcon from './components/SvgIcon'
-import Text from './components/Text'
-import Typography from './components/Typography'
-import VideoPlayer from './components/VideoPlayer'
-import XIcon from './components/XIcon'
-import Footer from './components/Footer/Footer'
-import FooterBrands from './components/Footer/FooterBrands'
-import FooterList from './components/Footer/FooterList'
-import SolutionForm from './components/SolutionForm'
+import ApplicantBadge from './components/ui/ApplicantBadge'
+import ApplicantListFilter from './components/forms/ApplicantListFilter'
+import Button from './components/ui/Button'
+import CollapsibleSection from './components/ui/CollapsibleSection'
+import ColorPalette from './components/ui/ColorPalette'
+import ErrorBox from './components/ui/ErrorBox'
+import HeaderBar from './components/ui/HeaderBar'
+import HorizontalRuler from './components/ui/HorizontalRuler'
+import Milestones from './components/ui/Milestones'
+import Roadmap from './components/ui/Roadmap'
+import RoadmapHero from './components/ui/RoadmapHero'
+import RoadmapLevel from './components/ui/RoadmapLevel'
+import RoadmapTimelineElement from './components/ui/RoadmapTimelineElement'
+import SvgIcon from './components/ui/SvgIcon'
+import Text from './components/ui/Text'
+import Typography from './components/ui/Typography'
+import VideoPlayer from './components/ui/VideoPlayer'
+import XIcon from './components/ui/XIcon'
+import Footer from './components/ui/Footer/Footer'
+import FooterBrands from './components/ui/Footer/FooterBrands'
+import FooterList from './components/ui/Footer/FooterList'
+import SolutionForm from './components/forms/SolutionForm'
 
 export {
   ApplicantBadge,
