+++ webpack.standalone.js
@@ -1,4 +1,5 @@
 const path = require('path')
+
 const base = require('./webpack.config')
 
 module.exports = {
